(browser-tests) Rename test data fixture
The fixture does not only contain user info anymore
