+++ apps/browser-tests/src/tests/callout-response-guest.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect } from "@playwright/test";
-import { signInAdmin } from "#fixtures/user-info.json";
+import { signInAdmin } from "#fixtures/testData.json";
 
 const name = "Guest"; // At least 5 characters long
 const email = "abc@xyz.com";
